requirements.txt
python-dotenv
langchain
langchain-text-splitters
langchain-community
bs4
langchain-anthropic
langchain-openai
langchain-core
neo4j
fastapi
python-multipart
uvicorn
faiss-cpu
pypdf
